requirements.txt
asgiref==3.6.0
azure-common==1.1.28
azure-communication-email==1.0.0b2
azure-core==1.26.3
azure-mgmt-core==1.3.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
